+++ klaw-gui/src/panels/setting.rs
@@ -643,6 +643,7 @@ impl SettingPanel {
                             DarkThemePreset::Frappe,
                             DarkThemePreset::Macchiato,
                             DarkThemePreset::Mocha,
+                            DarkThemePreset::Blackpink,
                         ] {
                             if ui
                                 .selectable_label(